Adding support for FA5 specific menu classes ("fas", "fab" etc.)
Using icons with FontAwesome 5 often requires to specify the type of icon as one of the following:
- fas -> Solid
- far -> Regular
- fal -> Light
- fad -> Duotone
- fab -> Brands
(see: https://fontawesome.com/how-to-use/on-the-web/referencing-icons/basic-use)

When adding these classes to your menu items they are not filtered and added to the span tag, rather they are left as a class in the li tag (WP menu-item), because they are not searched for. By changing the matching function, preg_grep, support for these 5 classes can be accomplished.